Move zoom from annotator to components
