+++ application/ui/src/components/zoom/hooks/use-panning.hook.ts
