+++ application/ui/src/components/zoom/hooks/use-sync-zoom.hook.ts
