+++ application/ui/src/components/zoom/hooks/use-wheel-panning.hook.ts
