+++ application/ui/src/components/zoom/utils.ts
