+++ application/ui/src/components/zoom/zoom-transform.tsx
