+++ application/ui/src/components/zoom/zoom.module.scss
