+++ application/ui/src/features/annotator/tools/segment-anything-tool/segment-anything-tool.component.tsx
@@ -7,11 +7,11 @@ import { PointerEvent, useEffect, useRef, useState } from 'react';
 
 import { clampPointBetweenImage } from '@geti/smart-tools/utils';
 
+import { useZoom } from '../../../../components/zoom/zoom.provider';
 import { AnnotationShapeWithLabels } from '../../annotations/annotation-shape-with-labels.component';
 import { MaskAnnotations } from '../../annotations/mask-annotations.component';
 import { AnnotatorLoading } from '../../annotator-loading.component';
 import { useAnnotator } from '../../providers/annotator-provider.component';
-import { useZoom } from '../../providers/zoom/zoom.provider';
 import { type Annotation, type Shape } from '../../types';
 import { SvgToolCanvas } from '../svg-tool-canvas.component';
 import { getRelativePoint, removeOffLimitPoints } from '../utils';
