+++ application/ui/src/features/prompt-sidebar/prompt-modes/visual-prompt/captured-image/captured-image-layout.component.tsx
@@ -4,8 +4,8 @@
  */
 
 import { Grid, minmax, View } from '@geti/ui';
-import { ZoomTransform } from 'src/features/annotator/providers/zoom/zoom-transform';
-import { ZoomProvider } from 'src/features/annotator/providers/zoom/zoom.provider';
+import { ZoomTransform } from 'src/components/zoom/zoom-transform';
+import { ZoomProvider } from 'src/components/zoom/zoom.provider';
 
 import { type CapturedImageType } from '../types';
 import { CapturedImageActions } from './captured-image-actions.component';
